Renaming SampleHandler folder to Samples
