+++ Diagnostics/CombineMaCh3Chains.cpp
@@ -3,7 +3,7 @@
 
 // MaCh3 includes
 #include "manager/manager.h"
-#include "SampleHandler/Structs.h"
+#include "Samples/Structs.h"
 
 _MaCh3_Safe_Include_Start_ //{
 // ROOT includes
