+++ ParameterHandler/ParameterHandlerGeneric.h
@@ -2,7 +2,7 @@
 
 // MaCh3 includes
 #include "ParameterHandler/ParameterHandlerBase.h"
-#include "SampleHandler/Structs.h"
+#include "Samples/Structs.h"
 
 /// @brief Class responsible for handling of systematic error parameters with different types defined in the config. Like spline, normalisation parameters etc.
 /// @see For more details, visit the [Wiki](https://github.com/mach3-software/MaCh3/wiki/02.-Implementation-of-Systematic).
