+++ ParameterHandler/ParameterHandlerUtils.h
@@ -1,7 +1,7 @@
 #pragma once
 
 // MaCh3 includes
-#include "SampleHandler/Structs.h"
+#include "Samples/Structs.h"
 
 _MaCh3_Safe_Include_Start_ //{
 // ROOT includes
