+++ Samples/SampleHandlerBase.cpp
