+++ Samples/SampleHandlerFD.cpp
@@ -1,5 +1,5 @@
 #include "SampleHandlerFD.h"
-#include "SampleHandler/Structs.h"
+#include "Samples/Structs.h"
 
 _MaCh3_Safe_Include_Start_ //{
 #include "Oscillator/OscillatorFactory.h"
