+++ mcmc/MCMCProcessor.h
@@ -10,7 +10,7 @@
 
 // MaCh3 includes
 #include "mcmc/StatisticalUtils.h"
-#include "SampleHandler/HistogramUtils.h"
+#include "Samples/HistogramUtils.h"
 
 _MaCh3_Safe_Include_Start_ //{
 // ROOT includes
